feat: Add PDF template type
Adds a new PDF type of template, accessible in the templates section for user officers.

It's a basic example for now which will require further refining and refactoring in the future to bring it more in line with other templates, which are all questionary-backed templates. Components have been re-used where possible but there is added complexity to re-using some components due to PDF templates being non-questionary.

Part of UserOfficeProject/user-office-project-issue-tracker#613